requirements.txt
Flask
mysql-connector-python
flask-sqlalchemy

